% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zoom.R
\name{check_zoom}
\alias{check_zoom}
\title{Internal check for a valid combination of CRS, map style and zoom level}
\usage{
check_zoom(
  zoom,
  style = c("outdoor", "road", "light", "leisure"),
  crs = c(27700, 3857),
  allow_premium = FALSE
)
}
\arguments{
\item{zoom}{Zoom level, an integer. For \code{crs = 27700}, this must be between
0 and 13. For \code{crs = 3857}, this must be between 7 and 20. Certain
combinations of zoom level and style retrieve tiles from the OS's Premium
service, which can generate a financial charge once past a certain level of
usage. See the OS Maps API webpages for details.}

\item{crs}{CRS code (EPSG), an integer, either \code{27700} (British National
Grid) or \code{3857} (standard web mapping projected coordinate system eg Google
Maps, OSM).}
}
\value{
No visible return unless an error is thrown
}
\description{
Internal check for a valid combination of CRS, map style and zoom level
}
