% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_bbox}
\alias{create_bbox}
\title{Use {osmdata} to easily get a bbox object for a given place name}
\usage{
create_bbox(place, crs)
}
\arguments{
\item{place}{string: place name to search for}

\item{crs}{integer or string: the EPSG code for the desired CRS}
}
\description{
Only the first search result from osmdata gets used, however
}
