% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_png.R
\name{generate_png_27700}
\alias{generate_png_27700}
\title{generate_png_27700}
\usage{
generate_png_27700(bbox, zoom, type, chatty = FALSE, ...)
}
\arguments{
\item{bbox}{A bbox object (a length 4 vector with names xmin, ymin, xmax, ymax)}

\item{zoom}{Zoom level, an integer}

\item{type}{Map type, a string. One of "Road", "Outdoor", "Light", "Leisure".}

\item{chatty}{Provide a brief info summary of the parameters that will be used. Requires the \code{usethis} package to be installed. Defaults to FALSE.}

\item{...}{params to pass to check_zoom(): allow_premium}
}
\description{
BNG: 700km (E|x) by 1300km (N|y) ie 7*13 = 91 100km^2 squares
}
\details{
Create a set of PNGs and their spatial extents
}
