% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_png.R
\name{generate_png_3857}
\alias{generate_png_3857}
\title{generate_png_3857}
\usage{
generate_png_3857(bbox, zoom, type, ...)
}
\arguments{
\item{bbox}{A bbox object (a length 4 vector with names xmin, ymin, xmax, ymax)}

\item{zoom}{Zoom level, an integer}

\item{type}{Map type, a string. One of "Road", "Outdoor", "Light", "Leisure".}

\item{...}{params to pass to check_zoom(): allow_premium and chatty}
}
\description{
Create a set of PNGs and their spatial extents
}
