% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_features_api.R
\name{query_features_api}
\alias{query_features_api}
\title{Retrieve data from the OS Features API}
\usage{
query_features_api(
  x,
  version = c("2.0.0", "1.1.0", "1.0.0"),
  user_agent = NULL
)
}
\arguments{
\item{x}{character. The name of a place, or a postcode for example. The API
docs say 'A free string text search of OS Names, intended to be an
ambiguous/fuzzy search.'}

\item{version}{character. API version. Most recent (2.0.0) by default.}

\item{user_agent}{character. \code{NULL} by default, which will use the package
name.}
}
\value{
A raw response from the API endpoint - you will want to process
this.
}
\description{
Function not yet completed and functional.
See https://osdatahub.os.uk/docs/wfs/technicalSpecification for details.
}
