% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_png.R
\name{query_maps_api}
\alias{query_maps_api}
\title{query_maps_api}
\usage{
query_maps_api(x, y, zoom, type, crs, user_agent = NULL)
}
\arguments{
\item{x}{Tile number (horizontally)}

\item{y}{Tile number (vertically)}

\item{zoom}{Zoom level, an integer}

\item{type}{Map type, a string. One of "Road", "Outdoor", "Light", "Leisure".}

\item{crs}{CRS code (EPSG), an integer, either 27700 or 3857}

\item{user_agent}{A User-Agent string to pass to the API}
}
\value{
Raw PNG data
}
\description{
Build and perform a query to the OS Maps API
}
