% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_names_api.R
\name{query_names_api}
\alias{query_names_api}
\title{Retrieve data from the OS Names API}
\usage{
query_names_api(
  x,
  local_types = NULL,
  bounds = NULL,
  max_results = 1L,
  user_agent = NULL
)
}
\arguments{
\item{x}{character. The name of a place, or a postcode for example. The API
docs say 'A free string text search of OS Names, intended to be an
ambiguous/fuzzy search.'}

\item{local_types}{character vector. Types of result to return. Examples are:
'City', 'Village'. See API docs for all options. \code{NULL} by default (= no
filtering by type).}

\item{bounds}{bbox. The bounding box of an area to limit search to. \code{NULL} by
default - no filtering by area. Can be supplied as a vector of 4 numerics,
in the following XMIN,YMIN,XMAX,YMAX form: c(414000, 114000, 414100, 114100)}

\item{max_results}{integer. How many results to return. Can be from 1-100.
Set to 1 by default.}

\item{user_agent}{character. \code{NULL} by default, which will use the package
name.}
}
\value{
An sfc geospatial tibble
}
\description{
Only uses 'find' method currently, not 'nearest' yet.
See https://osdatahub.os.uk/docs/names/technicalSpecification for details.
}
\examples{

query_names_api(c("Stroud", "Gloucester"))
}
