% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_basemap.R
\name{build_basemap}
\alias{build_basemap}
\title{Build a basemap from the Ordnance Survey Maps API}
\usage{
build_basemap(bbox, zoom, crs = c(27700, 3857), ...)
}
\arguments{
\item{bbox}{A bbox object (a length 4 vector with names xmin, ymin, xmax,
ymax)}

\item{zoom}{Zoom level, an integer. For \code{crs = 27700}, this must be between
0 and 13. For \code{crs = 3857}, this must be between 7 and 20. Certain
combinations of zoom level and style retrieve tiles from the OS's Premium
service, which can generate a financial charge once past a certain level of
usage. See the OS Maps API webpages for details.}

\item{crs}{CRS code (EPSG), an integer, either \code{27700} (British National
Grid) or \code{3857} (standard web mapping projected coordinate system eg Google
Maps, OSM).}

\item{...}{Other details to pass on to generate_png_data()}
}
\value{
A geospatially-referenced raster... hopefully.
}
\description{
Build a basemap from the Ordnance Survey Maps API
}
\examples{
oxf <- create_bbox("Oxford", 27700)
oxford_basemap <- build_basemap(oxf, zoom = 5, style = "road", crs = 27700)

tmap::tm_shape(oxford_basemap, raster.downsample = FALSE) +
  tmap::tm_rgb(max.value = 1)
}
