% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_basemap.R
\name{build_basemap}
\alias{build_basemap}
\title{build_basemap}
\usage{
build_basemap(
  bbox,
  zoom,
  type = c("outdoor", "road", "light", "leisure"),
  crs = c(27700, 3857),
  cache_tiles = FALSE,
  allow_premium = FALSE,
  chatty = TRUE
)
}
\arguments{
\item{bbox}{A bbox object (a length 4 vector with names xmin, ymin, xmax, ymax)}

\item{zoom}{Zoom level, an integer}

\item{type}{Map type, a string. One of "Road", "Outdoor", "Light", "Leisure".}

\item{crs}{CRS code (EPSG), an integer, either 27700 or 3857}

\item{cache_tiles}{Instead of converting PNG data from the API into a basemap on the fly, cache data as local PNG files. Not functional yet; defaults to FALSE.}

\item{allow_premium}{Whether to only access tiles from zoom levels that are within the "OpenData" tiers of the OS API. These vary according to map type and CRS. See https://osdatahub.os.uk/docs/wmts/technicalSpecification for details. Defaults to FALSE. Set to TRUE if you wish to access higher zoom levels within the "Premium" data tiers (chargeable - see https://osdatahub.os.uk/plans)}

\item{chatty}{Provide a brief info summary of the parameters that will be used. Requires the \code{usethis} package to be installed. Defaults to FALSE.}
}
\value{
A spatially-referenced raster, hopefully.
}
\description{
Build a basemap from the Ordnance Survey Maps API
}
