% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zoom.R
\name{check_zoom}
\alias{check_zoom}
\title{check_zoom}
\usage{
check_zoom(
  zoom,
  type = c("outdoor", "road", "light", "leisure"),
  crs = c(27700, 3857),
  free_only = TRUE,
  chatty = FALSE
)
}
\arguments{
\item{zoom}{zoom level, an integer}

\item{type}{map type, a string}

\item{crs}{CRS code (EPSG), an integer, either 27700 or 3857 (string is ok, arg will be coerced to string anyway)}

\item{free_only}{whether to only access tiles from zoom levels that are within the "OpenData" tiers of the OS API. These vary according to map type and CRS. See https://osdatahub.os.uk/docs/wmts/technicalSpecification for details. Defaults to TRUE. Set to FALSE if you wish to access higher zoom levels within the "Premium" data tiers (chargeable - see https://osdatahub.os.uk/plans)}

\item{chatty}{provide a brief info summary of the parameters that will be used. Requires the usethis package to be installed. Defaults to FALSE.}
}
\value{
no visible return unless an error is thrown
}
\description{
Check params are valid
}
