% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_zoom.R
\name{check_zoom}
\alias{check_zoom}
\title{Internal check for a valid combination of CRS, map style and zoom level}
\usage{
check_zoom(
  zoom,
  style = c("outdoor", "road", "light", "leisure"),
  crs = c(27700, 3857),
  allow_premium = FALSE,
  chatty = rlang::is_interactive()
)
}
\arguments{
\item{zoom}{Zoom level, an integer. For \code{crs = 27700}, this must be between
0 and 13. For \code{crs = 3857}, this must be between 7 and 20. Certain
combinations of zoom level and style retrieve tiles from the OS's Premium
service, which can generate a financial charge once past a certain level of
usage. See the OS Maps API webpages for details.}

\item{style}{Map style, a string. One of "road", "outdoor", "light",
"leisure". Leisure is only available in the 27700 CRS.}

\item{crs}{CRS code (EPSG), an integer, either \code{27700} (British National
Grid) or \code{3857} (standard web mapping projected coordinate system eg Google
Maps, OSM).}

\item{allow_premium}{Whether to only access tiles from zoom levels that are
within the "OpenData" tier of the OS API. These vary according to map style
and CRS. See the \href{https://osdatahub.os.uk/docs/wmts/technicalSpecification}{API Technical Specification} for details. Defaults to FALSE. Set to TRUE if
you wish to access zoom levels within the "Premium" service tier (chargeable
\itemize{
\item see \url{https://osdatahub.os.uk/plans})
}}

\item{chatty}{Whether to provide a summary of the parameters that will be
used, and the data and number of tiles that will be downloaded.
Requires the \code{usethis} package to be installed. By default, \code{FALSE} if not
in an interactive (\code{interactive()}) session, TRUE if in an interactive
session.}
}
\value{
No visible return unless an error is thrown
}
\description{
Internal check for a valid combination of CRS, map style and zoom level
}
