% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapirosa-package.R
\docType{package}
\name{mapirosa-package}
\alias{mapirosa}
\alias{mapirosa-package}
\title{mapirosa: Access OS Open Data APIs From R}
\description{
A package to enable accessing basemap tiles from the OS Maps API, initially
}
\author{
\strong{Maintainer}: Fran Barton \email{fbarton@alwaysdata.net}

}
\keyword{internal}
