% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_maps_api.R
\name{query_maps_api}
\alias{query_maps_api}
\title{Build and perform a query to the OS Maps API}
\usage{
query_maps_api(x, y, zoom, style, crs, user_agent = NULL)
}
\arguments{
\item{x}{Tile number (horizontally)}

\item{y}{Tile number (vertically)}

\item{zoom}{Zoom level, an integer. For \code{crs = 27700}, this must be between
0 and 13. For \code{crs = 3857}, this must be between 7 and 20. Certain
combinations of zoom level and style retrieve tiles from the OS's Premium
service, which can generate a financial charge once past a certain level of
usage. See the OS Maps API webpages for details.}

\item{crs}{CRS code (EPSG), an integer, either \code{27700} (British National
Grid) or \code{3857} (standard web mapping projected coordinate system eg Google
Maps, OSM).}

\item{user_agent}{A User-Agent string to pass to the API}
}
\value{
Raw PNG data
}
\description{
Build and perform a query to the OS Maps API
}
